ResumeParserJobRecommendation/requirements.txt
spacy
pdfminer.six
streamlit
python-docx
scikit-learn
pandas
